Fold type-cache isolation regression test into GradleParserTest
